+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 Changes
 =======
 
+## UNRELEASED
+
+* Bug: Handle top-level `await` in source files.
+
 ## 0.4.1
 
 * Bug: Handle `class` fields (public, private, static).